src/requirements.txt
azure-functions
markitdown[all]
